requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
anthropic>=0.34.0
google-generativeai>=0.8.0
requests>=2.31.0
pillow>=10.0.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
streamlit-local-storage>=0.0.5

# Optional but recommended for advanced lens corrections
# Note: lensfunpy requires system libraries (lensfun)
# On Ubuntu/Debian: sudo apt-get install liblensfun-dev
# On macOS: brew install lensfun
# Then: pip install lensfunpy
# lensfunpy>=1.13.0
scipy>=1.11.0  # Required for advanced image processing in lens corrections
opencv-python>=4.8.0  # For advanced image processing in Python agent
opencv-contrib-python>=4.8.0  # Additional OpenCV modules for white balance
Wand>=0.6.11  # Python bindings for ImageMagick